Introducing TableBody to better modularize Table component
