+++ kie-wb-common-react/boxed-expression-editor/src/components/Table/index.ts
@@ -16,6 +16,7 @@
 
 export * from "./EditableCell";
 export * from "./Table";
+export * from "./TableBody";
 export * from "./TableHandler";
 export * from "./TableHandlerMenu";
 export * from "./TableHeader";
